+++ be/src/exec/hdfs_scanner_orc.cpp
@@ -124,7 +124,8 @@ bool OrcRowReaderFilter::filterMinMax(size_t rowGroupIdx,
             if (row_idx_iter == rowIndexes.end()) {
                 return false;
             }
-            const orc::proto::ColumnStatistics& stats = row_idx_iter->second.entry(rowGroupIdx).statistics();
+            const orc::proto::ColumnStatistics& stats =
+                    row_idx_iter->second.entry(static_cast<int>(rowGroupIdx)).statistics();
             ColumnPtr min_col = min_chunk->columns()[i];
             ColumnPtr max_col = max_chunk->columns()[i];
             DCHECK(!min_col->is_constant() && !max_col->is_constant());
@@ -136,7 +137,7 @@ bool OrcRowReaderFilter::filterMinMax(size_t rowGroupIdx,
         } else {
             // search partition columns.
             int part_idx = 0;
-            const int part_size = _scanner_ctx.partition_columns.size();
+            const int part_size = static_cast<int>(_scanner_ctx.partition_columns.size());
             for (part_idx = 0; part_idx < part_size; part_idx++) {
                 if (_scanner_ctx.partition_columns[part_idx].col_name == slot->col_name()) {
                     break;
@@ -251,7 +252,7 @@ bool OrcRowReaderFilter::filterOnPickStringDictionary(
                 size_t old_size = offset_data[i + 1] - offset_data[i];
                 size_t new_size = remove_trailing_spaces(s, old_size);
                 bytes.insert(bytes.end(), s, s + new_size);
-                offsets[i] = total_size;
+                offsets[i] = static_cast<uint32_t>(total_size);
                 total_size += new_size;
             }
             offsets[dict_size] = total_size;
