+++ be/src/exprs/array_expr.cpp
@@ -62,7 +62,7 @@ public:
             for (const Column* element : element_raw_ptrs) {
                 array_elements->append(*element, i, 1);
             }
-            curr_offset += num_elements;
+            curr_offset += static_cast<uint32_t>(num_elements);
             array_offsets->append(curr_offset);
         }
